requirements.txt
nltk==3.2.5
GitPython==2.1.11
